% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchOrg.R
\name{BatchOrganize}
\alias{BatchOrganize}
\title{Organize the samples in folders (by batch)}
\usage{
BatchOrganize(metadata, remove)
}
\arguments{
\item{metadata}{path to a .csv file with the experiment metadata.}

\item{remove}{logical value (TRUE by default), if TRUE the files will be deleted after the sorting.}
}
\value{
Directories will be created (by batch) and the files will be placed according to the metadata information.
}
\description{
Put the spectra of the sample files into their respective batch folder. The .csv file must be the "name" and "batch" column.
The values in the column name need to have the same name of the sample files including the pattern (e.g.: .mzML, .CDF, .mzXML). Before perform this function, use the "setwd()" to the folder where the files are.
}
