% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_read.R
\name{chrom_find_data_start_line}
\alias{chrom_find_data_start_line}
\title{Find the Line Where Data Begins}
\usage{
chrom_find_data_start_line(file, n_lines = 50)
}
\arguments{
\item{file}{File path to the file to read.}

\item{n_lines}{Number of lines to search for the start of the data.}
}
\value{
Single integer of the start of the data.
}
\description{
Finds the line where the tabular data begins. This is then used to
start the reading of the data with `readr::read_csv()` and end the reading
of the metadata.
}
