% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_read.R
\name{chrom_interp_volume}
\alias{chrom_interp_volume}
\title{Interpolate Volume from Time}
\usage{
chrom_interp_volume(.data, time, volume)
}
\arguments{
\item{.data}{Data frame with a volume column to be interpolated.}

\item{time}{Name of the time column to use for interpolation.}

\item{volume}{Name of the volumn column to interpolate along the time.}
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Interpolates the volume column values, based on the time column values.
}
\details{
Some FPLC systems don't report accurate volume data, but do report accurate
timing data. This function takes the time and volume data and interpolates
the otherwise constant volume data in accordance with the time points.
}
\examples{
fl <- system.file(
  "extdata",
  "sec.txt",
  package = "chromr"
)

dat <- fl \%>\%
  chrom_read_quadtech(interp_volume = FALSE)

dat

dat \%>\%
  chrom_interp_volume(time, volume)

}
