% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_plot.R
\name{chrom_plot_fractions}
\alias{chrom_plot_fractions}
\title{Plot Chromatogram with Fractions}
\usage{
chrom_plot_fractions(data, wl_show = NULL, frac_wl = 280, frac_labelling = 5)
}
\arguments{
\item{data}{Datafram containing values.}

\item{wl_show}{Wavelengths to show on the plot.}

\item{frac_wl}{Wavelengths to show the fractionation scheme for.}

\item{frac_labelling}{How often to label the fraction scheme.}
}
\value{
a ggplot object.
}
\description{
Plot Chromatogram with Fractions
}
\examples{
fl1 <- system.file("extdata",
                   "20220809_SFPQfl_TEVdig_S200_part1.TXT",
                   package = "chromr")
fl2 <- system.file("extdata",
                   "20220809_SFPQfl_TEVdig_S200_part2.TXT",
                   package = "chromr")
df1 <- chrom_read_quadtech(fl1)
df2 <- chrom_read_quadtech(fl2)
dat <- chrom_append_run(df1, df)
chrom_plot_fractions(dat, frac_wl = c(280, 488))
}
