% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_read.R
\name{chrom_read}
\alias{chrom_read}
\title{Read chromatogram file.}
\usage{
chrom_read(file, ID = NULL, df = FALSE, tidy = TRUE, header = NULL)
}
\arguments{
\item{file}{Path to file.}

\item{ID}{String to identify the particular chromatogram / run.}

\item{df}{Return just dataframe, or list containing dataframe and other run details.}

\item{tidy}{Logical, whether or not to return a "tidy" tibble.}

\item{header}{Position of header values (NULL attempts to auto-detect).}
}
\value{

}
\description{
Read chromatogram file.
}
