% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_read.R
\name{chrom_read_ngc}
\alias{chrom_read_ngc}
\title{Read .csv Chromatogram from the BioRad NGC}
\usage{
chrom_read_ngc(file, skip = 1)
}
\arguments{
\item{file}{File path to the `.csv` file.}

\item{skip}{Number of lines to skip before begin reading. Usually 1 line.}
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Read .csv Chromatogram from the BioRad NGC
}
\examples{
fl <- system.file(
  "extdata",
  "ngc_sec.csv",
  package = "chromr"
)

dat <- chrom_read_ngc(fl)
dat

}
