% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_read.R
\name{chrom_read_quadtech}
\alias{chrom_read_quadtech}
\title{Read BioRad QuadTech Chromatogram Files}
\usage{
chrom_read_quadtech(file, interp_volume = TRUE)
}
\arguments{
\item{file}{Exported `.TXT` chromatogram file from the BioRad QuadTech.}

\item{interp_volume}{Logical. If TRUE, interpolates the values in the volume
column based on the values in the time column.}
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Read BioRad QuadTech Chromatogram Files
}
\examples{
fl <- system.file("extdata",
  "sec.txt",
  package = "chromr"
)
# just read
fl |>
  chrom_read_quadtech()

# read without interpolating volume
fl |>
  chrom_read_quadtech(interp_volume = FALSE)
# read then plot
fl |>
  chrom_read_quadtech() |>
  chrom_plot()
}
