% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromr-package.R
\docType{package}
\name{chromr-package}
\alias{chromr}
\alias{chromr-package}
\title{chromr: Read and Plot FPLC Chromatograms}
\description{
For reading files from BioRad QuadTec machines and making plots of the charomatograms.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BradyAJohnston/chromr}
  \item Report bugs at \url{https://github.com/BradyAJohnston/chromr/issues}
}

}
\keyword{internal}
