% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChromR.R
\docType{package}
\name{ChromR}
\alias{ChromR}
\title{ChromR package}
\description{
This package contains functions to help chromatographers in the diary tasks.
}
\author{
Estevan Bruginski \email{<estevan.bruginski@gmail.com>}
}
