% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{.compute_group_frac}
\alias{.compute_group_frac}
\title{Compute Group Statistics for Fraction Labels}
\usage{
.compute_group_frac(
  data,
  scales,
  period = 5,
  xpos = mean,
  ypos = 0,
  lines = NA,
  initial = 0L
)
}
\arguments{
\item{data}{Input data frame.}

\item{scales}{Scale objects from the plot.}

\item{period}{Integer specifying the interval between labels.}

\item{xpos}{Function or static value for x position.}

\item{ypos}{Function or static value for y position.}

\item{lines}{Vector of line colors to include.}

\item{initial}{Integer specifying the starting point.}
}
\value{
A data frame with computed positions and labels.
}
\description{
Internal function that computes the positions for fraction labels based on
specified grouping and positioning rules.
}
\keyword{internal}
