% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_akta_old.R
\name{.find_common_start}
\alias{.find_common_start}
\title{Find common start of multiple strings}
\usage{
.find_common_start(strings)
}
\arguments{
\item{strings}{Vector of strings to be processed}
}
\value{
Single character that is common to the start of all the strings.
 Returns empty string if nothing in common.
}
\description{
Walks along all strings until there is a character that is not shared
between all of them, then returns the string up until there.
}
\keyword{internal}
