% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatPeakList.R
\name{FormatPeakList}
\alias{FormatPeakList}
\title{Format Peak List}
\usage{
FormatPeakList(
  samplelist,
  peaklist,
  export = TRUE,
  expname = "BC_peaklist.csv"
)
}
\arguments{
\item{samplelist}{Datapath character, sample list in .csv with the columns names as follow: Samples, Class, Batch and, Order.}

\item{peaklist}{Datapath character, peaktable output from MetaboAnalystR (metaboanalyst_input.csv).}

\item{export}{Logical, if TRUE export the formated peaklist to a .csv file, TRUE by default.}

\item{expname}{Logical, plot the gradient profile, TRUE by default.}
}
\value{
Formated samplelist dataframe
}
\description{
Format and merge the sample list with the generated peak list.
}
\author{
Estevan Bruginski \email{estevan.bruginski@ufpr.br}, Letícia Meier \email{leticiameier@ufpr.br}

Universidade Federal do Paraná
License: MIT + file LICENSE
}
