% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\docType{data}
\name{grad}
\alias{grad}
\title{Example Data for gradient functions}
\format{An object of class \code{data.frame} with 6 rows and 4 columns.}
\usage{
grad
}
\description{
Example of gradient profile in a LC system.
}
\keyword{datasets}
