% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GradCalc.R
\name{GradCalc}
\alias{GradCalc}
\title{Gradient mobile phase calculation}
\usage{
GradCalc(x, runs, over = 15)
}
\arguments{
\item{x}{data frame, gradient information previously imported with GradInput function.}

\item{runs}{numeric, number of chromatographic runs in the batch.}

\item{over}{numeric, is an overage of mobile phase to keep the input of the pump submerged to avoid air entrance in the system. The default value is 15\%. Set according with your system.}
}
\value{
a list with the total batch time and the amount, in ml, of solvents A and B.
}
\description{
Calculate the A and B mobile phases of a LC gradient.
}
