% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gradient.R
\docType{data}
\name{gradient}
\alias{gradient}
\title{Gradient profile method of LC.}
\format{
An object of class \code{data.frame} with 6 rows and 4 columns.
}
\usage{
gradient
}
\description{
Example of gradient profile for LC system.
}
\examples{
\dontrun{
 gradient
}
}
\author{
Estevan Bruginski \email{estevan.bruginski@ufpr.br}
Universidade Federal do Paraná
License: MIT + file LICENSE
}
\keyword{datasets}
