% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GradInput.R
\name{GradInput}
\alias{GradInput}
\title{Import gradient}
\usage{
GradInput(path)
}
\arguments{
\item{path}{filepath to a .csv file.}
}
\value{
dataframe with the gradient information.
}
\description{
Import the gradient information from a .csv file.
}
