% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_metadata.R
\docType{data}
\name{metadata_ex}
\alias{metadata_ex}
\title{Metadata example.}
\format{
A data frame with 6 rows and 4 variables:
\describe{
\item{Samples}{Sample names.}
\item{Class}{Class identification.}
\item{Order}{Injection order.}
\item{Batch}{Batch identification.}
}
}
\usage{
metadata_ex
}
\description{
Example of metadata for FileOrganizer.
}
\examples{
\dontrun{
 metadata_ex
}
}
\author{
Estevan Bruginski \email{estevan.bruginski@ufpr.br}
Universidade Federal do Paraná
License: MIT + file LICENSE
}
\keyword{datasets}
