% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_akta_old.R
\name{read_akta_old}
\alias{read_akta_old}
\title{Read a .asc chromatogram from an old AKTA Pure}
\usage{
read_akta_old(file, fraction_size = 2)
}
\arguments{
\item{file}{Filepath to read}
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Read a .asc chromatogram from an old AKTA Pure
}
\examples{
fl <- system.file('extdata', 'example_akta_old.asc', package = 'wellr')

read_akta_old(fl)
}
