% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_columns}
\alias{rename_columns}
\title{Standardize Column Names}
\usage{
rename_columns(data)
}
\arguments{
\item{data}{A data frame whose columns need to be renamed}
}
\value{
A data frame with standardized column names
}
\description{
Renames specific columns in a data frame to standardized names.
Current mappings: volume -> Volume, time -> Time, fraction -> Fraction
}
\examples{
\dontrun{
df <- data.frame(volume = 1:3, time = 4:6, other = 7:9)
renamed_df <- rename_columns(df)  # 'volume' becomes 'Volume', 'time' becomes 'Time'
}
}
