% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\docType{data}
\name{StatFraction}
\alias{StatFraction}
\title{Fraction Bar Stat for ggplot2}
\format{
A ggproto object extending 'Stat'
}
\usage{
StatFraction
}
\description{
ggproto object defining the statistical transformation for fraction bars.
}
\keyword{internal}
