% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\docType{data}
\name{StatFractionGroup}
\alias{StatFractionGroup}
\title{Fraction Label Stat for ggplot2}
\format{
A ggproto object extending 'Stat'
}
\usage{
StatFractionGroup
}
\description{
ggproto object defining the statistical transformation for fraction labels.
}
\keyword{internal}
