% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcGrad.R
\name{CalcGrad}
\alias{CalcGrad}
\title{Cálculo do volume de fase móvel e tempo total de um batch cromatográfico}
\usage{
CalcGrad(x, runs, over = 15)
}
\arguments{
\item{x}{Datapath character, arquivo em .csv do gradiente.}

\item{runs}{Numeric, número de corridas no batch.}

\item{over}{Numeric, limite de segurança para o sistema cromatográfico não secar. Valor padrão é 15%.}
}
\value{
uma lista com o volume total de cada fase móvel em mL (A e B) e tempo total
}
\description{
Cálculo do volume de fase móvel e tempo total de um batch cromatográfico.
}
\examples{
\dontrun{
GradCalc(x, runs = 15, over = 15)
}
}
\author{
Estevan Bruginski \email{estevan.bruginski@ufpr.br}
Universidade Federal do Paraná
License: MIT + file LICENSE
}
