% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_column_exist}
\alias{check_column_exist}
\title{Check if Column is Present in DataFrame Ignoring Case}
\usage{
check_column_exist(data, name)
}
\arguments{
\item{data}{Dataframe}

\item{name}{string of column name}
}
\value{
Logical TRUE/FALSE if the column exists.
}
\description{
Check if Column is Present in DataFrame Ignoring Case
}
