% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.R
\name{chrom_append_run}
\alias{chrom_append_run}
\title{Append a Chromatogram to Another Run}
\usage{
chrom_append_run(.data, ..., adjust_fractions = FALSE)
}
\arguments{
\item{.data}{A data frame created through `chrom_read_*()`}

\item{...}{Runs created through `chrom_read_*()`}

\item{adjust_fractions}{Logical. FALSE keeps the fraction numbering as-is.
TRUE starts the number of the subsequent fractions by adding the previous
fractions.}
}
\value{
a [tibble][tibble::tibble-package] which combines the given runs.
}
\description{
Sums the volume and the time columns. Optionally continues the numbering of
the fractions after the previous runs.
}
\examples{

fl1 <- system.file("extdata",
                   "20220809_SFPQfl_TEVdig_S200_part1.TXT",
                   package = "chromr")
fl2 <- system.file("extdata",
                   "20220809_SFPQfl_TEVdig_S200_part2.TXT",
                   package = "chromr")
df1 <- chrom_read_quadtech(fl1)
df2 <- chrom_read_quadtech(fl2)


df1 \%>\%
  chrom_append_run(df2)
}
