% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_read.R
\name{chrom_interp_volume}
\alias{chrom_interp_volume}
\title{Interpolate Volume from Time}
\usage{
chrom_interp_volume(df, time, volume)
}
\arguments{
\item{df}{}
}
\description{
Interpolates the volume column values, based on the time column values.
}
\details{
Some FPLC systems don't report accurate volume data, but do report accurate
timing data. This function takes the time and volume data and interpolates
the otherwise constant volume data in accordance with the time points.
}
