% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_plot.R
\name{chrom_plot}
\alias{chrom_plot}
\title{Create basic plot of chromatogram.}
\usage{
chrom_plot(data, channels = NULL, xlim = NULL, ylim = NULL)
}
\arguments{
\item{channels}{Character vector of valengths to plot (e.g. \code{c("280", "260")}).}

\item{chrom}{Chrom object from \code{read_chrom()}.}
}
\description{
Create basic plot of chromatogram.
}
