% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_read.R
\name{chrom_read_quadtech}
\alias{chrom_read_quadtech}
\title{Read BioRad QuadTech Chromatogram Files}
\usage{
chrom_read_quadtech(file, interp_volume = TRUE)
}
\arguments{
\item{file}{Exported `.TXT` chromatogram file from the BioRad QuadTech.}

\item{interp_volume}{Logical. If TRUE, interpolates the values in the volume column based on the values in the time column.}
}
\description{
Read BioRad QuadTech Chromatogram Files
}
