% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChromR.R
\docType{package}
\name{ChromR}
\alias{ChromR}
\title{ChromR package}
\description{
The ChromR package brings some functions to help analysts, Ph.D. students and technicians to plan and organize LC analyses.
The functions are separated into 2 main sections, Pre-analyse and Post-analyse. In Pre-analyse, the functions will help to randomize
the injections order,  calculation of the quantity of mobile phase and the spent time of batches.
The Post-analyse functions will help to organize the obtained files according to their respective batches or classes,
avoiding human error.
}
\author{
Estevan Bruginski \email{<estevan.bruginski@gmail.com>}
}
