% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interpolate_column}
\alias{interpolate_column}
\title{Interpolate Interpolate Given Column}
\usage{
interpolate_column(data, col)
}
\arguments{
\item{data}{Dataframe with column to interpolate.}

\item{col}{Name of the column to interpolate.}
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Interpolate Interpolate Given Column
}
