% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGrad.R
\name{PlotGrad}
\alias{PlotGrad}
\title{Plot Gradiente profile}
\usage{
PlotGrad(grad)
}
\arguments{
\item{grad}{data frame, gradient information previously imported with GradInput function.}
}
\value{
The graph of gradient profile.
}
\description{
Plot the LC gradient profile.
}
