% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_columns}
\alias{rename_columns}
\title{Title}
\usage{
rename_columns(data)
}
\arguments{
\item{data}{Dataframe to rename the columns of.}
}
\value{
Dataframe with renamed column.
}
\description{
Title
}
