% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleRand.R
\name{SampleRand}
\alias{SampleRand}
\title{Plot Gradiente profile}
\usage{
SampleRand(x, gnumber, gnames)
}
\arguments{
\item{x}{data frame, the first column is the id of the samples and the second column is the groups of each sample.}

\item{gnumber}{numeric, number of groups in the sample, need to be between 2 and 6.}

\item{gnames}{character, names of the groups, need to be the same of the input data frame.}
}
\value{
A randomized sample list.
}
\description{
Plot the LC gradient profile.
}
\examples{
\dontrun{
SampleRand(x, gnumber = 3, gnames = c("g1", "g2", "g3"))
}
}
\author{
Estevan Bruginski \email{estevan.bruginski@ufpr.br}
Universidade Federal do Paraná
License: MIT + file LICENSE
}
